move conversion module into dedicated utils
